moving UndoHandle, second attempt
