+++ Bio/File.py
@@ -7,14 +7,9 @@
 # package.
 """Code for more fancy file handles.
 
-Classes:
- - UndoHandle     File object decorator with support for undo-like operations.
-
-Additional private classes used in Bio.SeqIO and Bio.SearchIO for indexing
-files are also defined under Bio.File but these are not intended for direct
-use.
+Bio.File defines private classes used in Bio.SeqIO and Bio.SearchIO for
+indexing files. These are not intended for direct use.
 """
-
 
 import codecs
 import os
@@ -153,7 +148,7 @@ def _open_for_random_access(filename):
 
 
 class UndoHandle:
-    """A Python handle that adds functionality for saving lines.
+    """A Python handle that adds functionality for saving lines (DEPRECATED).
 
     Saves lines in a LIFO fashion.
 
@@ -165,94 +160,10 @@ class UndoHandle:
 
     def __init__(self, handle):
         """Initialize the class."""
-        self._handle = handle
-        self._saved = []
-        try:
-            # If wrapping an online handle, this this is nice to have:
-            self.url = handle.url
-        except AttributeError:
-            pass
-
-    def __iter__(self):
-        """Iterate over the lines in the File."""
-        return self
-
-    def __next__(self):
-        """Return the next line."""
-        next = self.readline()
-        if not next:
-            raise StopIteration
-        return next
-
-    def readlines(self, *args, **keywds):
-        """Read all the lines from the file as a list of strings."""
-        lines = self._saved + self._handle.readlines(*args, **keywds)
-        self._saved = []
-        return lines
-
-    def readline(self, *args, **keywds):
-        """Read the next line from the file as string."""
-        if self._saved:
-            line = self._saved.pop(0)
-        else:
-            line = self._handle.readline(*args, **keywds)
-        return line
-
-    def read(self, size=-1):
-        """Read the File."""
-        if size == -1:
-            saved = "".join(self._saved)
-            self._saved[:] = []
-        else:
-            saved = ""
-            while size > 0 and self._saved:
-                if len(self._saved[0]) <= size:
-                    size = size - len(self._saved[0])
-                    saved = saved + self._saved.pop(0)
-                else:
-                    saved = saved + self._saved[0][:size]
-                    self._saved[0] = self._saved[0][size:]
-                    size = 0
-        return saved + self._handle.read(size)
-
-    def saveline(self, line):
-        """Store a line in the cache memory for later use.
-
-        This acts to undo a readline, reflecting the name of the class: UndoHandle.
-        """
-        if line:
-            self._saved = [line] + self._saved
-
-    def peekline(self):
-        """Return the next line in the file, but do not move forward though the file."""
-        if self._saved:
-            line = self._saved[0]
-        else:
-            line = self._handle.readline()
-            self.saveline(line)
-        return line
-
-    def tell(self):
-        """Return the current position of the file read/write pointer within the File."""
-        return self._handle.tell() - sum(len(line) for line in self._saved)
-
-    def seek(self, *args):
-        """Set the current position at the offset specified."""
-        self._saved = []
-        self._handle.seek(*args)
-
-    def __getattr__(self, attr):
-        """Return File attribute."""
-        return getattr(self._handle, attr)
-
-    def __enter__(self):
-        """Call special method when opening the file using a with-statement."""
-        return self
-
-    def __exit__(self, type, value, traceback):
-        """Call special method when closing the file using a with-statement."""
-        self._handle.close()
-
+        raise Exception("The UndoHandle class has been deprecated, and was "
+                        "moved to Bio/SearchIO/_legacy/ParserSupport.py "
+                        "(which is the only module in Biopython still using "
+                        "UndoHandle.")
 
 # The rest of this file defines code used in Bio.SeqIO and Bio.SearchIO
 # for indexing
