+++ Tests/test_File.py
@@ -9,65 +9,10 @@ import os.path
 import shutil
 import tempfile
 import unittest
+from io import StringIO
 
 from Bio import bgzf
 from Bio import File
-from io import StringIO
-
-data = """This
-is
-a multi-line
-file"""
-
-
-class UndoHandleTests(unittest.TestCase):
-    """Tests for the UndoHandle."""
-
-    def test_one(self):
-        """First test."""
-        h = File.UndoHandle(StringIO(data))
-        self.assertEqual(h.readline(), "This\n")
-        self.assertEqual(h.peekline(), "is\n")
-        self.assertEqual(h.readline(), "is\n")
-        # TODO - Meaning of saveline lacking \n?
-        h.saveline("saved\n")
-        self.assertEqual(h.peekline(), "saved\n")
-        h.saveline("another\n")
-        self.assertEqual(h.readline(), "another\n")
-        self.assertEqual(h.readline(), "saved\n")
-        # Test readlines after saveline
-        h.saveline("saved again\n")
-        lines = h.readlines()
-        self.assertEqual(len(lines), 3)
-        self.assertEqual(lines[0], "saved again\n")
-        self.assertEqual(lines[1], "a multi-line\n")
-        self.assertEqual(lines[2], "file")  # no trailing \n
-        # should be empty now
-        self.assertEqual(h.readline(), "")
-        h.saveline("save after empty\n")
-        self.assertEqual(h.readline(), "save after empty\n")
-        self.assertEqual(h.readline(), "")
-
-    def test_read(self):
-        """Test read method."""
-        h = File.UndoHandle(StringIO("some text"))
-        h.saveline("more text")
-        self.assertEqual(h.read(), "more textsome text")
-
-    def test_undohandle_read_block(self):
-        """Test reading in blocks."""
-        for block in [1, 2, 10]:
-            s = StringIO(data)
-            h = File.UndoHandle(s)
-            h.peekline()
-            new = ""
-            while True:
-                tmp = h.read(block)
-                if not tmp:
-                    break
-                new += tmp
-            self.assertEqual(data, new)
-            h.close()
 
 
 class RandomAccess(unittest.TestCase):
